requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
Flask-Testing
python-dotenv
